backend/requirements.txt
# Core API
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.22
python-dotenv==1.2.1
starlette==0.37.2

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Auth & Security
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
cryptography==46.0.4
ecdsa==0.19.1
PyJWT==2.11.0

# AI Providers
openai==1.99.9
google-generativeai==0.8.6
google-genai==1.62.0
google-ai-generativelanguage==0.6.15
litellm==1.80.0
tiktoken==0.12.0

# Google APIs
google-api-python-client==2.189.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
google-api-core==2.29.0
proto-plus==1.27.1
protobuf==5.29.6

# HTTP
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
aiohttp==3.13.3
aiohttp-retry==2.9.1

# WebSockets / Realtime
python-socketio==5.16.1
python-engineio==4.13.1
simple-websocket==1.1.0
websockets==15.0.1
bidict==0.23.1
wsproto==1.3.2

# Twilio
twilio==9.10.2

# Payments
stripe==14.3.0

# Email
resend==2.21.0

# AWS / S3
boto3==1.42.42
botocore==1.42.42
s3transfer==0.16.0

# Data
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
email-validator==2.3.0
python-dateutil==2.9.0.post0
pandas==3.0.0
numpy==2.4.2

# PDF Generation (pure Python, no system dependencies)
fpdf2==2.8.7
defusedxml==0.7.1

# Validation
jsonschema==4.26.0
referencing==0.37.0
jsonschema-specifications==2025.9.1
rpds-py==0.30.0

# Async
anyio==4.12.1
sniffio==1.3.1
uvloop==0.22.1
watchfiles==1.1.1
httptools==0.7.1

# Utils
PyYAML==6.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
Markdown==3.10.2
click==8.3.1
rich==14.3.2
tenacity==9.1.2
regex==2026.1.15
six==1.17.0
packaging==26.0
psutil==7.2.2
pillow==12.1.0
lxml==6.0.2
fastuuid==0.14.0
jq==1.11.0

# Office docs (only if actually used)
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5
xlsxwriter==3.2.9

# Certificates / HTTP
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
httplib2==0.31.2

# gRPC (required by google-api-core)
grpcio==1.76.0
grpcio-status==1.71.2

# Misc dependencies
tzdata==2025.3
typing_extensions==4.15.0
multidict==6.7.1
yarl==1.22.0
frozenlist==1.8.0
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
propcache==0.4.1
attrs==25.4.0
oauthlib==3.3.1
requests-oauthlib==2.0.0
rsa==4.9.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
cachetools==5.5.2
